+++ integrations/google-analytics/HISTORY.md
@@ -1,7 +1,6 @@
-2.18.1 / 2020-02-07
+2.18.2  / 2020-02-18
 ===================
-
-  * Add support for all casing types for enhanced ecommerce checkout options.
+  * Fix a bug where we are sending a product level category for the event category. We will send Google Analytics an event category of `'EnhancedEcommerce'` for the following events: `'product clicked'`, `product removed`, `product added` and `product viewed`.
 
 
 2.18.0 / 2019-12-17